+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-## []
+## [1.1.1]
 
 - Removed entity tooltip in REI
 - Fixed enchantment level coloring